make config reading relative to cwd
Function `relativeConfigFile` indicates that config `require`-ing expected to be cwd-relative. But in real that is not true (bare `require` was used, so it is module-relative).

In this case it is not really possible to import config relative to the dir the "binary" running.
For now workaround is to use absolute path.
```
CFG=`pwd`/cfg/database.json

./node_modules/.bin/sequelize --config ${CFG} --env dev --migrate
```
Instead of simple:
```
./node_modules/.bin/sequelize --config ./cfg/database.json --env dev --migrate
```

Pull request fixes this issue.