Fix bug where setting same date twice would prevent persistence
This issue was caused by the difference in comparisons between
Instance.set and Instance.changed. Dates were being compared with !==
in #set, but with #valueOf in #changed. The _previousDataValue
would thus always get overwritten when setting a date field. This means
that performing a set on a Date field twice with two different Date
objects with the *same* #valueOf would trick Sequelize into thinking
that the field was clean.