should fix sequelize/sequelize#3284
adjusts the query for insert and update for mssql when triggers are
being used.  It adds a temp table, which will hold the inserted/updated
row and then select it from there.  Before this, it was retrieving the
changed row by select Inserted.* which didn’t work if triggers were on
the table.