Fix findOrCreate logic. Closes #2970
There is a BC here: If defaults is set, it will only use the defaults object. This is necessary to be sure what the user actually wants. Otherwise there will always be some problems by combining complex where with the defaults.

Only one test is breaking by this change. Other than that one issue is fixed and from now on it's possible to use complex where statements.

I also inserted a update lock as a default

Update docs