Check if ENUM value exists before ALTER TYPE
Currently sequelize will not check if an ENUM value exists before attempting to alter the type. If it does it can throw an error such as:

```
SequelizeDatabaseError: enum label "[labelName]" already exists
      at [object Object].Query.formatError (node_modules/sequelize/lib/dialects/postgres/query.js:433:14)
      at [object Object].<anonymous> (node_modules/sequelize/lib/dialects/postgres/query.js:108:19)
      at [object Object].Query.handleError (node_modules/pg/lib/query.js:108:8)
      at [object Object].<anonymous> (node_modules/pg/lib/client.js:171:26)
      at Socket.<anonymous> (node_modules/pg/lib/connection.js:109:12)
      at readableAddChunk (_stream_readable.js:146:16)
      at Socket.Readable.push (_stream_readable.js:110:10)
      at TCP.onread (net.js:523:20)
```

By adding the `IF NOT EXISTS` qualifier it will only throw a `NOTICE` that the enum label exists and skip it.

Check for Postgres 9.3.0 and greater

require in semver

Add check for 9.3.0 ENUM feature

Remove extra quotation mark

Break out conditional SQL statement into IF block