fix model validation documentation
In 4.0, `model.validate()` was changed to return a promise that resolves
only on success and rejects on error. Previously, it would return a
promise that resolved to either `undefined` on success or some
`ValidationError` on error. The upgrade guide did not mention this
breaking change, and the documentation did not reflect the change. A
test was added to show the intent of the documentation.