fix(connection): fixing mssql reconnect bug
There is currently a bug where if you restart a mssql database,
sequelize will lock the pool resources.  Once those are locked, all
future requests to .acquire will throw a ResourceRequest timed out
error.  This commit wraps all the connect logic in a retry to prevent an
inifite loop of trying to connect, and also watchins for an 'end' event
in the mssql adapter to properly reject the promise if 'end' is sent
before 'connect' is fired.