Add feature issues to exempt issues closing
I'm not sure if this is desired or not, but I added the `feature` label to the exempt issues labels.  I did this because it seem to me that closing a PR due to inactivity alone isn't a great experience for the community. 

For example the streaming issue here #2454 that likely needs another issue completed to be finished #2325 are both marked as `stale` and were closed. 

From a user standpoint, and a person who wants both of those features, it's disappointing to see them get closed without explanation with each has a good amount of interest in the Sequelize community.