Handle passing in undefined as an assoc target
When using the patterns demonstrated here: https://github.com/sequelize/express-example/blob/master/models/index.js, it's possible to accidentally pass in `undefined` as the target of an association. This results in a stack trace that isn't obvious. Instead, the lines that assume that the target is defined can check that assumption.