Add option to `sync` similar to `alter` but without `removeColumn`
This is a 2-line change that I thought would be useful. The idea is that it would allow the dev to write something like `User.sync({alterKeepColumns: true})` which does the same thing as `{alterKeepColumns: true}` only it doesn't remove old columns. This way you don't have to risk data deletion if you were to typo or something.