test(isolation): should be able to switch isolation levels
Added failing tests for MySQL and MariaDB.

There is currently no way to check they isolation level of a transaction (@@transaction_isolation tells you what the next transaction in the session will use, but does nothing about the current transaction), so checking it directly in an integration test is not possible. Instead these tests check that specifying an isolation level results in the expected behaviour for that isolation level.