fix(docs): fix package id links in docs
These were previously being rendered via esdoc (which uses marked's
autolink rendering) as `mailto:` links.  There are some
[suggestions](https://gist.github.com/alexpeattie/4729247) about how one
might disable or avoid autolinking but in this case the package
identifiers are in backticks, producing weird behaviour similar to
https://github.com/markedjs/marked/issues/1327.

Assuming the fix attached to this latter issue applied in this case,
then a dependency upgrade of `marked` within `esdoc` might solve the
problem, but there are [maintenance
issues](https://github.com/esdoc/esdoc/pull/543) there.

Until that arrives, this is a workaround to avoiding the weird `mailto:`
links on this page.