feat(postgres): change returning option to only return model attributes
If an underlying table has columns that don't exist in the model,
when using `returning: true` the instance will be created with these
extra attributes. This change alters this behavior to only return the
fields defined in the model, provided a model definition is passed.

BREAKING CHANGE: setting `returning: true` will no longer create
attributes on the instance that are not defined in the model.

The old default behavior, if desired, can be used by changing the:

`returning: true`

to

`returning: ['*']`