docs(manual/other-topics/hooks): Fix broken link to `hooks.js`
The current manual link points to `sequelize/lib/hooks.js` which was probably changed during a project file structure update.
For more inexperienced people who are looking to learn more about the available hooks, looking through the source code is the only option. When their main entryopint is the manual, they're greeted with a broken link.