fix(upsert): fall back to DO NOTHING if no update key values provided
We found that if we did an upsert on a record with the `fields` option, if the data values being passed into the upsert did not include any of those fields, that we'd get a postgres syntax error when the query was executed.

example:
```
   // for this example, MyModel has attributes id, a, b, x, and y.
   const [ instance ] = await MyModel.upsert({ id: 1, x: 1, y, 2 }, { fields: [ 'a', 'b' ], returning: true });
   // this would result in the following query:
   //    INSERT INTO "MyModels" ("id", "x", "y")
   //      VALUES ($1, $2, $3)
   //       ON CONFLICT ("id") DO UPDATE SET RETURNING "id", "a", "b", "x", "y";
   //                                      ^^^ notice the missing clause here.
```

If there are no update values included, and we call `upsert`, I would expect the query to still attempt the insert and just not update anything if there's a conflict.

To fix, we added code into the abstract query generator that will recognize that none of the update values were actually provided, and revert to use a `DO NOTHING` instead of a `DO UPDATE`.

For the `ON DUPLICATE KEY UPDATE` sql languages, we try fall back to just updating the identity column that triggered the duplicate constraint (per https://stackoverflow.com/questions/4596390/insert-on-duplicate-key-do-nothing).  If there are no identity columns nor update values, then we throw an error.

I'm also fixing an error in the mssql upsertQuery, where it was looking at the `updateValues` instead of the `insertValues` to determine if the `IDENTITY_INSERT` needed to be turned on.  This was preventing my unit test for the above fix from passing.