refactor(build): use npm script to run commands with docker env vars
NPM package.json options are not widely known and may be confusing. Moreover, in NPM v7 the `options` dictionarry is renamed to `config` and breaks the current functionality https://docs.npmjs.com/cli/v7/configuring-npm/package-json#config.
By re-using an npm script to run commands with docker-specific environment variables we make the configuration more understandable and compatible with the current and the upcomming node versions