refactor(typescript): add foundation for incremential TS migration
Converts the logger interface to TypeScript.
Adds eslint rules and TypeScript configuration to allow for a gradual migration to TS.
Adds a build script to transpile both TS and JS code while keeping the same API for package usage.
Changes imports from test files to use "absolute" paths that point to the directory of the compiled code.