feat(postgresql): easier SSL config and options param support
This commit uses the pg_connection_string package to parse the
connection string if the dialect is postgresql. This is helpful because
it automatically handles reading SSL certs that are specified in the
connection string.

As part of this, support was added for the `options` URL parameter,
which allows arbitrary session variables to be configured in the
connection string.