feat(postgres): support for adding foreign keys with NOT VALID
This extends `addConstraint` and `changeColumn` to support
adding foreign keys (via `ALTER` statements) with a `NOT VALID`
option. It does it by introducing a new attribute `notValid: true`
on the references object (`ModelAttributeColumnReferencesOptions` type).

This is specific to postgres dialect only. Added unit and integration specs,
and update type definition.