docs(contributing-cheatsheet): add new cheatsheet docs
Cheatsheet is an extenstion to CONTRIBUTING.md. This flavor focuses on the PostgreSQL dialect.

It covers:
- Sequelize tests
  - prequisite steps
  - running various tests and expectations
  - debugging expectations
- using docker (basic commands)
  - creating docker networks
  - inspecting docker containers
  - connecting containers to docker networks
- setting up pgadmin4
  - running as a container
  - connecting it with database container
  - configuring pgadmin settings

This is not intended as a replacement to CONTRIBUTING.md, which contains must-read information. This document is intended as a quick guide targeting those new to Docker and Sequelize development. There is great opportunity for stylistic and content improvements.