feat: improved error reporting on associations (#14901)
* feat: improved error reporting on associations
This improves the error reporting on associations by telling us what exactly is
not a proper subclass of the Sequelize Model. This assists in debugging and
should present little risk to the existing behavior.

Possibly due to #14161 I was not able to get the full test suite running
locally _before_ I made any local changes. However, the number of tests that
were failing (2) remained the same after my changes, and the failing tests had
more to do with SQLite than changes to the model structure.

* Apply suggestions from code review

Co-authored-by: Rik Smale <13023439+WikiRik@users.noreply.github.com>